Fixed types

+++ oghma/src/main/scala/dev/pb/oghma/writer/OghmaMarshal.scala
@@ -5,12 +5,12 @@ import dev.pb.oghma.api.Marshal
 import dev.pb.oghma.common.{BufferUtils, OghmaTag}
 
 sealed trait OghmaMarshal[F[_]] extends Marshal[F]:
-  def marshalInt(value: Int): Stream[F, Byte] =
+  def marshalInt(value: Long): Stream[F, Byte] =
     value match
       case x if x >= -128 && x <= 127 => marshalInt8(x)
 
-  private def marshalInt8(value: Int): Stream[F, Byte] =
-    val buffer = BufferUtils.bigEndianByteBuffer(OghmaTag.Int8.value, value)
+  private def marshalInt8(value: Long): Stream[F, Byte] =
+    val buffer = BufferUtils.bigEndianByteBuffer(OghmaTag.Int8.value, value.toByte)
     BufferUtils.toStream(buffer)
 
 object OghmaMarshal:
